Refactor http adapter
