--- lib/pow_assent/http_response.ex
@@ -1,12 +1,0 @@
-defmodule PowAssent.HTTPResponse do
-  @moduledoc false
-
-  @type header :: {binary(), binary()}
-  @type t      :: %__MODULE__{
-    status: integer(),
-    headers: [header()],
-    body: binary()
-  }
-
-  defstruct status: 200, headers: [], body: ""
-end
